+++ src/ansys/fluent/core/docker/docker_compose.py
@@ -1,6 +1,6 @@
 """Launch Fluent through docker compose."""
 
-# Copyright (C) 2021 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2021 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #