+++ src/ansys/fluent/core/services/datamodel_se.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2021 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2021 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -1648,12 +1648,9 @@ class PyNamedObjectContainer:
 
     def get_object_names(self) -> Any:
         """Displays the name of objects within a container."""
-        if self.service.version <= FluentVersion.v241:
-            return self._get_child_object_display_names()
-        else:
-            return self.service.get_object_names(
-                self.rules, convert_path_to_se_path(self.path)
-            )
+        return self.service.get_object_names(
+            self.rules, convert_path_to_se_path(self.path)
+        )
 
     getChildObjectDisplayNames = get_object_names