+++ src/ansys/fluent/core/services/field_data.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2021 - 2025 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2021 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -993,7 +993,7 @@ class Batch(FieldBatch):
             Data is returned as dictionary of dictionaries in the following structure:
             tag int | Tuple-> surface_id [int] -> field_name [str] -> field_data[np.array]
 
-            The tag is a tuple for Fluent 2023 R1 or later.
+            The tag is a tuple.
         """
         return BatchFieldData(
             ChunkParser().extract_fields(